Update startup-troubleshooting.md
On ArcoLinux KDE, i can't start LBRY App anymore, it says "Error Starting Up", and that's it. I have tryed to delete the config directory, reinstalling the App, but none of this helps. Soo please, look at this bug, it is anoying that i can't use the App anymore. Thanks, and hope to solve this error soon.